Deleted mistakes in interface

+++ src/ru/ioffe/school/susanin/mapGraph/Edge.java
@@ -5,7 +5,7 @@ package ru.ioffe.school.susanin.mapGraph;
  */
 interface Edge {
 
-    int getTime();
+    double getTime();
     int getCost();
     Vertex getFrom();
     Vertex getTo();
